+++ src/main/java/com/spring_boot/webshop_app/service/ItemService.java
@@ -2,6 +2,7 @@ package com.spring_boot.webshop_app.service;
 
 import com.spring_boot.webshop_app.dto.ItemDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,7 +16,7 @@ public interface ItemService {
 
     List<ItemDto> findAllInPriceRange(Long uprLmt, Long lwrLimit);
     @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
-    List<ItemDto> filter(Optional<Integer[]> ids, Optional<Long> uprLmt, Optional<Long> lwrLmt,
+    List<ItemDto> filter(Optional<Integer[]> ids, Optional<BigDecimal> uprLmt, Optional<BigDecimal> lwrLmt,
                          Optional<Integer> productTypeId, Optional<Integer> productionYear,
                          Optional<String> sortBy, Optional<String> sortOrder);
